Defer GDI+ initialization to first PInvoke access (#11781)
Initializing GDI+ on module load left GDI+ with a stale DPI setting. It has to come after the call to Application.SetHighDpiMode.

Fixes 11608

There is one case where we can call a GDI+ PInvoke in Core before we call any in System.Drawing (creating a Graphics object from HWND).

I've moved the initialization down and explicitly validated initialization in this case and in one other possible location in Core where we might use GDI+ before we've ever called something in System.Drawing's interop.
