+++ src/System.Drawing.Common/src/System/Drawing/ToolboxBitmapAttribute.cs
@@ -311,9 +311,6 @@ public class ToolboxBitmapAttribute : Attribute
 #pragma warning disable CA1810 // DummyFunction apparently needs to be invoked prior to the rest of the initialization
     static ToolboxBitmapAttribute()
     {
-        // When we call Gdip.DummyFunction, JIT will make sure Gdip..cctor will be called.
-        Gdip.DummyFunction();
-
         Stream? stream = BitmapSelector.GetResourceStream(typeof(ToolboxBitmapAttribute), "DefaultComponent.bmp");
         Debug.Assert(stream is not null, "DefaultComponent.bmp must be present as an embedded resource.");
