+++ src/System.Private.Windows.Core/src/NativeMethods.txt
@@ -73,6 +73,8 @@ GdipGetImagePixelFormat
 GdipGetImageRawFormat
 GdipGetRegionHRgn
 GdipIsInfiniteRegion
+GdiplusStartup
+GdiplusStartupInputEx
 GdipSaveImageToStream
 GetClientRect
 GetClipRgn
