+++ src/System.Private.Windows.Core/src/System/Drawing/ImageCodecInfoHelper.cs
@@ -36,6 +36,8 @@ internal static class ImageCodecInfoHelper
         {
             if (s_encoders is null)
             {
+                GdiPlusInitialization.EnsureInitialized();
+
                 uint numEncoders;
                 uint size;
