+++ src/System.Private.Windows.Core/src/Windows/Win32/Graphics/GdiPlus/GdiplusStartupInputEx.cs
